A tad more minor cleanup, contemplating what to do with the concept of executables.
Moved some old module files into the deprecated folder.
Deleted some now defunct launching methods as there will now be no need
to tell steam to launch starbound directly.
Changed the layout of the application config json so executable are a
vector with a name field per executable rather than a map of names to
executable information.
Contemplating how to handle executables as a filesystem resource. One
possible method will be to just let the user define executables manually
and keep the current structure pretty much as-is. Another method would
be to automatically handle downloading and unpacking executables for
both osb and xsb. For pratical reasons, I will probably go with the
former and potentially squeeze the latter in when ready.