Move submenus into separate module, inputs and messages for selecting files to create executables.
Moved the Settings and CreateProfile submenus into a 'menus' module.
Added a text-input field for the name, and two buttons that file-pick
the binary and .pak files of an executable.
The "Add" button cannot be clicked unless a name and path is provided.
No executable is "created" yes, but a CreateExecutable message is now
available in the Application Message structure.