Temporary fix for loading assets
Osb was not loading assets, turns out you can't specify pak files
directly. This hack just gets the parent of the pak file. This may be
correct behavior, it may not. Possibly selecting "folders" plural of pak
files could be better behavior.
Xsb is refusing to load atm due to missing dependencies (SDL2 ??), so I
will be work on that another time.