Update README to reflect current status of project
With the addition of GUI-based executable creation, instructions for
manually editing the config.json is no longer necessary. All the user
has to do now is just unzip the executables into a reasonable directory.
Also move the project goal of "Downloading and Updating executables..."
to a lower and optional priority, this may not be within the bounds of
this project and be better supported as a different project entirely.