Reference Executables using the key type rather than a usize
The settings menu used to reference the Executables in a map by index
rather than key. This means that the key had to be derived from an index
that may not have been stable to iterate against. The proper things has
been done and the reference is done by storing a String key rather than
an integer.