Removing default selected executable sets default to None, fixed some warnings.
When an executable is removed, if the executable was also the default
executable, it would continue to be saved in the config until a new
executable was selected. This did not impact the functionality of the
code, as SBI already handled the case of loading an invalid executable,
but fixing this was essentially a free operation to implement as the
config file was saved either way on the removal of an executable.
Some warnings related to unused variables was fixed relating to the file
names of selected binaries and assets.