Code cleanup, add new link mods dir bool to profile json
Writing the existing config to disk has been a replicated task over
multiple messages, so I created a function to generate the task and
replace the duplicated code.
Found a weird instance where a match statement was destructuring a
struct (I didnt even know you could do that), so I replaced it with an
actual destructuring statement.
Adding a new boolean to link the vanilla mods folder to the current
instance (requested feature, still yet to be used/implemented).