Implement functionality to link vanilla mods directory to profile.
If the link_mods value is true for the launching profile, the directory
of the vanilla mods folder is added to the assets list in the
sbinit.config.
Because the addition of the vanilla mods directory is now being managed
by the link_mods value, the weird way the default profile was adding the
mods folder as part of the "additional_mods" list was removed.
The path to the mods directory is also now stored separately in the Dirs
struct instead of needing to be derived.