Update README
The method of how to add an executable has changed slightly in that you
create the executable before configuring it.
Also realized that some of the lines ran really long, so I formatted the
file to keep lines from running offscreen on a 1080p monitor. I might
come back later and format all the text to not go beyond 80 cells.